Fix review bugs: TIN u8 overflow, genebody exonic counting, Qualimap secondary
- Fix TIN slot_idx u8 truncation: widen to u16 to support sample_size > 255
  without silent data corruption from index aliasing
- Fix genebody exonic/intronic base counting to use actual overlap length
  instead of full block length when blocks span exon-intron boundaries;
  cap exonic_bases at block_len to handle overlapping merged exons
- Fix Qualimap 'secondary' count: use actual BamStat secondary+supplementary
  values instead of computing as residual (which included QC-fail, paired-
  status-filtered reads, producing incorrect 'reads aligned' in Qualimap)
- Remove unused _interner parameter from process_chromosome_batch
- Add rust-version = "1.87" to Cargo.toml (required for is_multiple_of)

+++ src/rna/rseqc/tin.rs
@@ -53,7 +53,7 @@ pub struct TinIndex {
     pub transcripts: Vec<TranscriptSampling>,
     /// Per-chromosome sorted position entries for binary search.
     /// Each entry: (genomic_position, transcript_index, slot_index).
-    pub chrom_positions: HashMap<String, Vec<(u64, u32, u8)>>,
+    pub chrom_positions: HashMap<String, Vec<(u64, u32, u16)>>,
     /// Per-chromosome transcript spans for overlap pre-filtering.
     /// Each entry: (tx_start, tx_end, transcript_index), sorted by tx_start.
     pub chrom_spans: HashMap<String, Vec<(u64, u64, u32)>>,
@@ -240,7 +240,7 @@ impl TinIndex {
             self.chrom_positions
                 .entry(chrom_upper.clone())
                 .or_default()
-                .push((pos, tx_idx, slot_idx as u8));
+                .push((pos, tx_idx, slot_idx as u16));
         }
 
         // Add transcript span
@@ -325,7 +325,7 @@ pub struct TinAccum {
     pub read_starts: Vec<u32>,
     /// Number of sampled slots per transcript.
     #[allow(dead_code)]
-    pub n_samples: Vec<u8>,
+    pub n_samples: Vec<u16>,
     /// Minimum MAPQ threshold.
     pub mapq_cut: u8,
     /// Minimum coverage threshold.
@@ -342,7 +342,7 @@ impl TinAccum {
         for tx in &index.transcripts {
             let n = tx.sampled_positions.len();
             coverage.push(vec![0u32; n]);
-            n_samples.push(n as u8);
+            n_samples.push(n as u16);
         }
 
         TinAccum {
